feat: 모임 생성 Step3 컴포넌트 생성
  - Step3가 추가됨에 따라 기존의 Step3~7을 Step4~8로 변경
