+++ frontend/src/components/Create/Step3/index.tsx
@@ -1,68 +1,11 @@
-import { memo, LegacyRef, forwardRef } from 'react';
+import { forwardRef, LegacyRef } from 'react';
 
-import { DurationDate } from 'types/data';
-import { getNewDateString } from 'utils/date';
+import { Container } from '../@shared/styled';
 
-import {
-  Container,
-  ErrorColor,
-  Heading,
-  LabelContainer,
-} from '../@shared/styled';
-import * as S from './index.styled';
+interface Step3Props {}
 
-interface Step3Props {
-  useDateState: () => DurationDate & {
-    setStartDate: (startDate: DurationDate['startDate']) => void;
-    setEndDate: (endDate: DurationDate['endDate']) => void;
-  };
-  pressEnterToNext: (e: React.KeyboardEvent<HTMLInputElement>) => void;
+function Step3({}: Step3Props, ref: LegacyRef<HTMLDivElement>) {
+  return <Container ref={ref}>Step3</Container>;
 }
 
-function Step3(
-  { useDateState, pressEnterToNext }: Step3Props,
-  ref: LegacyRef<HTMLDivElement>,
-) {
-  const { startDate, setStartDate, endDate, setEndDate } = useDateState();
-
-  const changeStartDate = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setStartDate(e.target.value);
-
-    if (!endDate) {
-      setEndDate(e.target.value);
-    }
-  };
-
-  const changeEndDate = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setEndDate(e.target.value);
-  };
-
-  return (
-    <Container ref={ref}>
-      <Heading>
-        <span>언제부터 언제까지</span> 만날건가요? <ErrorColor>*</ErrorColor>
-      </Heading>
-      <LabelContainer>
-        <p>기간</p>
-        <S.InputWrapper>
-          <S.Input
-            type="date"
-            value={startDate}
-            onChange={changeStartDate}
-            min={getNewDateString('day')}
-          />
-          ~
-          <S.Input
-            type="date"
-            value={endDate}
-            onChange={changeEndDate}
-            min={startDate || getNewDateString('day')}
-            onKeyPress={pressEnterToNext}
-          />
-        </S.InputWrapper>
-      </LabelContainer>
-    </Container>
-  );
-}
-
-export default memo(forwardRef(Step3));
+export default forwardRef(Step3);
