+++ frontend/src/components/Create/Step5/index.tsx
@@ -1,53 +1,47 @@
 import { forwardRef, LegacyRef, memo } from 'react';
 
-import { convertRemainTime, getNewDateString } from 'utils/date';
+import { Container, Heading } from '../@shared/styled';
+import * as S from './index.styled';
 
-import {
-  Container,
-  ErrorColor,
-  Heading,
-  Input,
-  LabelContainer,
-  Label,
-} from '../@shared/styled';
+// TODO: 달력은 나중에 ^^
 
-interface Step5Props {
-  useDeadlineState: () => {
-    deadline: string;
-    setDeadline: (deadline: string) => void;
-  };
+interface Step4Props {
   pressEnterToNext: (e: React.KeyboardEvent<HTMLInputElement>) => void;
 }
 
-function Step5(
-  { useDeadlineState, pressEnterToNext }: Step5Props,
+function Step4(
+  { pressEnterToNext }: Step4Props,
   ref: LegacyRef<HTMLDivElement>,
 ) {
-  const { deadline, setDeadline } = useDeadlineState();
-  const changeDeadline = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setDeadline(e.target.value);
-  };
-
   return (
     <Container ref={ref}>
       <Heading>
-        <span>언제까지</span> 모집할까요? <ErrorColor>*</ErrorColor>
+        <span>언제</span> 만날건가요?
       </Heading>
-      <LabelContainer>
-        <Label>
-          <p>날짜</p>
-          <p>{deadline && `${convertRemainTime(deadline)} 후`}</p>
-        </Label>
-        <Input
-          type="datetime-local"
-          value={deadline}
-          onChange={changeDeadline}
-          onKeyPress={pressEnterToNext}
-          min={getNewDateString('min')}
-        />
-      </LabelContainer>
+      <S.Content>
+        <S.Calendar />
+        <S.Right>
+          <S.DailyButton type="button">매일</S.DailyButton>
+          <S.DayBox>
+            <span className="sun">일</span>
+            <span>월</span>
+            <span>화</span>
+            <span>수</span>
+            <span>목</span>
+            <span>금</span>
+            <span className="sat">토</span>
+          </S.DayBox>
+          <S.InputWrapper>
+            <S.Input type="time" />
+            부터
+            <S.Input type="time" onKeyPress={pressEnterToNext} />
+            까지
+          </S.InputWrapper>
+          <S.AddButton type="button">달력에 추가하기</S.AddButton>
+        </S.Right>
+      </S.Content>
     </Container>
   );
 }
 
-export default memo(forwardRef(Step5));
+export default memo(forwardRef(Step4));
