--- frontend/src/components/Create/Step6/index.styled.ts
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-import { LabelContainer as Label } from '../@shared/styled';
-
-const LabelContainer = styled(Label)`
-  max-width: 43.75rem;
-`;
-
-export { LabelContainer };
