+++ frontend/src/components/Create/index.ts
@@ -5,3 +5,4 @@ export { default as Step4 } from './Step4';
 export { default as Step5 } from './Step5';
 export { default as Step6 } from './Step6';
 export { default as Step7 } from './Step7';
+export { default as Step8 } from './Step8';
