+++ frontend/src/pages/Create/index.tsx
@@ -11,6 +11,7 @@ import {
   Step5,
   Step6,
   Step7,
+  Step8,
 } from 'components/Create';
 import Navigator from 'components/Create/Navigator';
 import { BROWSER_PATH } from 'constants/path';
@@ -23,11 +24,12 @@ import validator from './validate';
 const totalPage = [
   { number: 1, content: '이름 입력' },
   { number: 2, content: '카테고리 선택' },
-  { number: 3, content: '진행 날짜 선택' },
-  { number: 4, content: '날짜, 시간대 상세 입력' },
-  { number: 5, content: '모집 마감일자 입력' },
-  { number: 6, content: '장소 입력' },
-  { number: 7, content: '상세 설명 입력' },
+  { number: 3, content: '최대 인원 입력' },
+  { number: 4, content: '진행 날짜 선택' },
+  { number: 5, content: '날짜, 시간대 상세 입력' },
+  { number: 6, content: '모집 마감일자 입력' },
+  { number: 7, content: '장소 입력' },
+  { number: 8, content: '상세 설명 입력' },
 ];
 
 function Create() {
@@ -92,9 +94,8 @@ function Create() {
     }
 
     requestCreateGroup(groupData)
-      .then(res => {
-        const id = 1;
-        navigate(`${BROWSER_PATH.DETAIL}/${id}`);
+      .then(groupId => {
+        navigate(`${BROWSER_PATH.DETAIL}/${groupId}`);
       })
       .catch(error => {
         alert(error.message);
@@ -114,23 +115,24 @@ function Create() {
           ref={getPageRef(2)}
           gotoAdjacentPage={gotoAdjacentPage}
         />
-        <Step3
+        <Step3 ref={getPageRef(3)} />
+        <Step4
           useDateState={useDateState}
-          ref={getPageRef(3)}
+          ref={getPageRef(4)}
           pressEnterToNext={pressEnterToNext}
         />
-        <Step4 ref={getPageRef(4)} pressEnterToNext={pressEnterToNext} />
-        <Step5
-          useDeadlineState={useDeadlineState}
-          ref={getPageRef(5)}
-          pressEnterToNext={pressEnterToNext}
-        />
+        <Step5 ref={getPageRef(5)} pressEnterToNext={pressEnterToNext} />
         <Step6
-          useLocationState={useLocationState}
+          useDeadlineState={useDeadlineState}
           ref={getPageRef(6)}
           pressEnterToNext={pressEnterToNext}
         />
-        <Step7 useDescriptionState={useDescriptionState} ref={getPageRef(7)} />
+        <Step7
+          useLocationState={useLocationState}
+          ref={getPageRef(7)}
+          pressEnterToNext={pressEnterToNext}
+        />
+        <Step8 useDescriptionState={useDescriptionState} ref={getPageRef(8)} />
       </S.ScrollContainer>
       <Navigator
         page={page}
